Finished overview section
Summary: Added a section below the featured area that has 2, 3-coumn
rows of icons that explain what Rethink Mercy does. I am prettified the
code that is already there so it’s looking better and more orderly.
Next is to add a section of featured posts and begin testing the
offcanvas navigation menu.

Testing plan: N/A