Created events/what's on section, recompression
Summary: I added a section to the bottom of the page that also fits the
viewport.

Issues: I need to find a way to fix the overlap on the home screen
between the first and second sectionals

I also need to fix the sizing of the speaker images as well as reformat
the spacing so that it is as follows:
Image, hr, par. with quote, it. description of them
Testing plan: